--- .gitignore
@@ -1,3 +1,0 @@
-example/
-out.txt
-test.rb
